SDK SceneComposer : all shortcuts tool are now outside of the selectTool,  - added the Delete and Duplicate shortcut  - deleted the MoveManager  - clean up the SelectTool (removed old shortCuts)  - now the ShortcutManager can also provide to SorthcutTools the state of ctr, alt and shift key
